Uploaded Tuple.java and TupleDesc.java
I have successfully completed the code of the two files. I have completed the entire coding myself, though Manish helped me in the debugging part.
It didn't take long for me to code the Tuple file. I could finish it in a few hours. TupleDesc file too didn't take much time, except for the Iterator code. I took a lot of time in finding out the code. But in the end, I could do it.