Modified HeapPage.java and HeapFile.java
I have implementer the InsertTuple and the DeleteTuple methods. The code didn't take much of my efforts and I finished them in an hour or so. I felt a small trouble in the deleteTuple method of HeapPage.java, but I finished it too. I have also implemented writePage method in HeapPage.